pyproject.toml
[project]
name = "api-batch-tester"
version = "0.1.0"
description = "批量测试生图/生视频 API 的命令行工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.13.0",
    "pyyaml>=6.0",
    "tqdm>=4.67.3",
]
